ref: split ReceiveExecutor tests
